refactor: 모델 선택 컴포넌트 분리
