+++ src/pages/CarBuild/CarBuild.types.ts
@@ -1,0 +1,14 @@
+export interface CarModel {
+  id: number;
+  imgSrc: string;
+  name: string;
+  price: number;
+  model: string;
+  description: string;
+  isBest: boolean;
+  engineType: string;
+  options: {
+    defaultOptions: string[];
+    selectOptions?: string[];
+  };
+}
